Fix RSSI bugs

+++ Safety/Src/RSSI.cpp
@@ -6,7 +6,7 @@ static volatile int16_t counter = 0;
 
 //Time is in microseconds
 
-constexpr uint8_t MIN_ON_TIME = 1200; //20% of pwm signal from 1000 to 2000 microseconds
+constexpr uint16_t MIN_ON_TIME = 1200; //20% of pwm signal from 1000 to 2000 microseconds
 
 bool CommsFailed() {
     return failed;
